perf: optimize shuffle gather with sorted indices and reduced allocations
Three targeted optimizations to the existing deferred-gather shuffle
write path:

1. Sort partition indices by (batch_id, row_id) before gather. Arrow's
   interleave_record_batch has an internal optimization that coalesces
   adjacent rows from the same source batch into a single extend() call.
   Previously indices were in hash order (random batch_ids), defeating
   this optimization.

2. Skip BatchCoalescer for full-size batches. The gather path already
   produces batch_size-row batches via PartitionedBatchIterator, but
   these were cloned and pushed through the coalescer (a no-op for
   full-size batches). Now batches >= batch_size bypass the coalescer
   and write directly to the IPC buffer.

3. Eliminate (u32,u32) -> (usize,usize) index conversion. Store
   partition indices as (usize, usize) from the start, and borrow
   slices directly in PartitionedBatchIterator instead of allocating
   a converted Vec. This eliminates a 1.6MB allocation per partition
   during shuffle_write.

+++ native/core/src/execution/shuffle/partitioners/multi_partition.rs
@@ -110,7 +110,7 @@ pub(crate) struct MultiPartitionShuffleRepartitioner {
     output_data_file: String,
     output_index_file: String,
     buffered_batches: Vec<RecordBatch>,
-    partition_indices: Vec<Vec<(u32, u32)>>,
+    partition_indices: Vec<Vec<(usize, usize)>>,
     partition_writers: Vec<PartitionWriter>,
     shuffle_block_writer: ShuffleBlockWriter,
     /// Partitioning scheme to use
@@ -181,6 +181,7 @@ impl MultiPartitionShuffleRepartitioner {
             output_index_file,
             buffered_batches: vec![],
             partition_indices: vec![vec![]; num_output_partitions],
+
             partition_writers,
             shuffle_block_writer,
             partitioning,
@@ -422,7 +423,7 @@ impl MultiPartitionShuffleRepartitioner {
             let before_size = indices.allocated_size();
             indices.reserve(row_indices.len());
             for row_idx in row_indices {
-                indices.push((buffered_partition_idx, *row_idx));
+                indices.push((buffered_partition_idx as usize, *row_idx as usize));
             }
             let after_size = indices.allocated_size();
             mem_growth += after_size.saturating_sub(before_size);
